Added control panel listing view
