+++ src/bika/lims/controlpanel/configure.zcml
@@ -69,14 +69,6 @@
     />
 
     <browser:page
-      for="bika.lims.interfaces.ISampleConditions"
-      name="folder_view"
-      class=".bika_sampleconditions.SampleConditionsView"
-      permission="senaite.core.permissions.ManageBika"
-      layer="bika.lims.interfaces.IBikaLIMS"
-    />
-
-    <browser:page
       for="bika.lims.interfaces.ISampleMatrices"
       name="folder_view"
       class=".bika_samplematrices.SampleMatricesView"
