+++ src/senaite/core/browser/controlpanel/configure.zcml
@@ -24,6 +24,7 @@
   <include package=".instrumentlocations" />
   <include package=".interpretationtemplates" />
   <include package=".labels" />
+  <include package=".sampleconditions" />
   <include package=".samplecontainers" />
 
 </configure>
